requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Faker==16.6.1
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
numpy==1.24.2
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-env==0.8.1
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.0
sqlalchemy-model-factory==0.4.5
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
zipp==3.12.0
